+++ test/Signals/SignalCatchingOperationsTestCase.hpp
@@ -2,7 +2,7 @@
  * \file
  * \brief SignalCatchingOperationsTestCase class header
  *
- * \author Copyright (C) 2015 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ * \author Copyright (C) 2015-2017 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
  *
  * \par License
  * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
@@ -33,6 +33,15 @@ class SignalCatchingOperationsTestCase : public SignalsTestCaseCommon
 	constexpr static uint8_t testCasePriority_ {UINT8_MAX / 2};
 
 public:
+
+	/**
+	 * \return priority at which this test case should be executed
+	 */
+
+	constexpr static uint8_t getTestCasePriority()
+	{
+		return testCasePriority_;
+	}
 
 	/**
 	 * \brief SignalCatchingOperationsTestCase's constructor